+++ obelisk-server/src/main/java/org/burrow_studios/obelisk/server/moderation/db/ticket/TicketDB.java
@@ -2,23 +2,12 @@ package org.burrow_studios.obelisk.server.moderation.db.ticket;
 
 import com.google.gson.JsonObject;
 import org.burrow_studios.obelisk.server.db.DatabaseException;
-import org.burrow_studios.obelisk.server.moderation.ModerationService;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Set;
 
 public interface TicketDB {
-    static @NotNull TicketDB get(@NotNull ModerationService modService) {
-        final String host     = "null";
-        final int    port     = 3306;
-        final String database = "null";
-        final String user     = "null";
-        final String pass     = "null";
-
-        return new MySQLTicketDB(host, port, database, user, pass);
-    }
-
     @NotNull Set<Long> getTicketIds() throws DatabaseException;
 
     @NotNull JsonObject getTicket(long id) throws DatabaseException;
